core/requirements.txt
boto==2.49.0
boto3==1.10.50
botocore==1.13.50
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
gensim==3.8.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
numpy==1.18.1
pyemd==0.5.1
pystempel==1.1.0
python-dateutil==2.8.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.4.1
six==1.13.0
smart-open==1.9.0
sortedcontainers==2.1.0
tqdm==4.41.1
urllib3==1.25.7
Werkzeug==0.16.0
